refactoring: EditProfileModal를 Modals 폴더로 이동
