+++ frontend/guest-app/src/components/SideMenu/SideMenuBody.js
@@ -2,7 +2,7 @@ import React from "react";
 import List from "@material-ui/core/List";
 import Divider from "@material-ui/core/Divider";
 import SideMenuItem from "./SideMenuItem.js";
-import EditProfileModal from "../Modals/EditPriofileModal/EditProfileModal.js";
+import EditProfileModal from "../Modals/EditProfileModal.js";
 import useSideMenuStyles from "./UseSideMenuStyles.js";
 import {LogoutIcon, QuestionIcon, UserIcon} from "../FontAwesomeIcons.js";
 import useCommonModal from "../CommonModal/useCommonModal.js";
